requirements.txt
Django~=5.1.4
djangorestframework~=3.15.2
pip~=23.2.1
attrs~=24.3.0
pytz~=2024.2
PyYAML~=6.0.2
asgiref~=3.8.1
sqlparse~=0.5.3
swagger-spec-validator~=3.0.4
packaging~=24.2
inflection~=0.5.1
uritemplate~=4.1.1
referencing~=0.35.1
jsonschema~=4.23.0
jsonschema-specifications~=2024.10.1
setuptools~=65.5.0
typing_extensions~=4.12.2
importlib_resources~=6.5.2
drf-yasg~=1.21.8